fix storage

+++ server/controllers/files.js
@@ -39,24 +39,26 @@ Files.prototype.raw = function(ctx) {
 
 Files.prototype.token = async function(ctx) {
 	const key = decodeURIComponent(ctx.params.id);
-	const username = ctx.state.user.username;
+	const userId = ctx.state.user.userId;
 
-	const keys = key.split("/");
+	//const keys = key.split("/");
 	
-	if (keys.length < 2 || key.indexOf(username) != 0) return ERR.ERR_PARAMS();
-	const filetype = keys[0].replace(username+ "_", "");
+	//if (keys.length < 2 || key.indexOf(username) != 0) return ERR.ERR_PARAMS();
+	//const filetype = keys[0].replace(username+ "_", "");
 
-	//console.log(keys, key, filetype);
-	if (filetype != "pages" && filetype != "files" && filetype != "images" && filetype == "datas") {
-		return ERR.ERR_PARAMS();
-	}
+	////console.log(keys, key, filetype);
+	//if (filetype != "pages" && filetype != "files" && filetype != "images" && filetype == "datas") {
+		//return ERR.ERR_PARAMS();
+	//}
 
-	if (await storage.isFull(username)) {
+	if (await storage.isFull(userId)) {
 		return ERR.ERR().setMessage("存贮空间不足");
 	}
 
 	const result =  storage.getUploadToken(key);
 	if (result.isErr()) return result;
+
+	await this.model.upsert({userId, key});
 
 	const token = result.getData();
 	
@@ -65,83 +67,53 @@ Files.prototype.token = async function(ctx) {
 
 
 Files.prototype.statistics = async function(ctx) {
-	const username = ctx.state.user.username;
+	const userId = ctx.state.user.userId;
 
-	let data = await storage.getStatistics({username: username});
+	let data = await storage.getStatistics(userId);
 
 	return ERR.ERR_OK(data);
 }
 
 Files.prototype.upsert = async function(ctx) {
 	const params = ctx.state.params;
-	const key = decodeURIComponent(ctx.params.id);
-	const username = ctx.state.user.username;
-	params.username = username;
-	params.key = key;
-
-	if (key.indexOf(username + "/") != 0 && key.indexOf(username + "_") != 0) {
-		return ERR.ERR_PARAMS();
-	}
-	
-	const content = params.content;
-	params.content = undefined;
-	// 内容存在，则写文件
-	if (content) {
-		let result = await storage.upload(key, content);
-		if (result.isErr()) return result;
-
-		params.hash = result.getData().hash;
-	}
+	const userId = ctx.state.user.userId;
+	params.userId = userId;
 
 	let data = await this.model.upsert(params);
 
-	await storage.updateStatistics(username);
+	await storage.updateStatistics(userId);
 
 	return ERR.ERR_OK(data);
 }
 
 Files.prototype.delete = async function(ctx) {
-	const key = decodeURIComponent(ctx.params.id);
-	const username = ctx.state.user.username;
+	const id = ctx.params.id;
+	const userId = ctx.state.user.userId;
+	const where = {id, userId};
 
-	let data = await storage.delete(key);
+	let data = await this.model.findOne({where});
+	if (!data) return ERR.ERR_OK();
+
+	data = data.get({plain:true});
+	const key = data.key;
+
+	data = await storage.delete(key);
 	if (data.isErr()) return data;
 
-	data = await this.model.destroy({
-		where: {
-			key: key,
-			username: username,
-		}
-	});
-
-	await storage.updateStatistics(username);
+	data = await this.model.destroy({where});
+	await storage.updateStatistics(userId);
 
 	return ERR.ERR_OK(data);
 }
 
 Files.prototype.find = async function(ctx) {
 	const params = ctx.state.params;
-	const username = ctx.state.user.username;
-
-	if (params.raw) {
-		return storage.list(params.prefix || username);
-	}
-
+	const userId = ctx.state.user.userId;
+	const where = {userId};
 	
-	const where = {};
-	if (username) {
-		where.username = username;
-	} else {
-		where.public = true;
-	}
+	if (params.type) where.type = params.type;
 
-	if (params.type) {
-		where.type = params.type;
-	} else {
-		where.type = {[ne]:"pages"};
-	}
-
-	where.username = username || params.username;
+	where.size = {[gt]:0};
 
 	let data = await this.model.findAll({
 		where: where,
@@ -153,75 +125,40 @@ Files.prototype.find = async function(ctx) {
 }
 
 Files.prototype.findOne = async function(ctx) {
-	const key = decodeURIComponent(ctx.params.id);
-	const username = ctx.state.user.username;
-
-	let content = undefined;
-	let isPage = false;
-	if (_.endsWith(key, ".md")) {
-		content = await storage.get(key).then(ret => ret.getData());
-		isPage = true;
-	}
+	const id = ctx.params.id;
+	const userId = ctx.state.user.userId;
 
 	let data = await this.model.findOne({
 		where: {
-			key: key,
-			username: username,
+			id:id,
+			userId: userId,
 		}
 	})
-
-	if (!data) {
-		if (isPage) return ERR.ERR_OK({key, content});
-
-		return ERR.ERR_NOT_FOUND();
-	}
-
-	data = data.get({plain: true});
-	data.content = content;
 
 	return ERR.ERR_OK(data);
 }
 
 // 重命名
-Files.prototype.rename = async function(ctx) {
-	const filename = ctx.state.params.filename;
-	const id = ctx.params.id;
-	const username = ctx.state.user.username;
+//Files.prototype.rename = async function(ctx) {
+	//const filename = ctx.state.params.filename;
+	//const id = ctx.params.id;
+	//const userId = ctx.state.user.userId;
 	
-	let data = await this.model.findOne({
-		where: {
-			id: id,
-			username: username,
-		}
-	});
-
-	if (!data) return ERR.ERR_NOT_FOUND();
-
-	data = data.get({plain:true});
+	//data = await this.model.update({
+		//filename: filename,
+	//}, {
+		//where: {
+			//id: id,
+			//userId: userId,
+		//}
+	//})
 	
-	// 七牛改名
-	//const srcKey = username + "_files/" + data.filename;
-	//const dstKey = username + "_files/" + filename;
-	//data = await storage.move(srcKey, dstKey);
-	//if (data.isErr()) return data;
-	
-	data = await this.model.update({
-		filename: filename,
-	}, {
-		where: {
-			id: id,
-		}
-	})
-	
-	return ERR.ERR_OK(data);
-}
+	//return ERR.ERR_OK(data);
+//}
 
 Files.prototype.qiniu = async function(ctx) {
 	const params = ctx.request.body;
-	const {key} = params;
-	const username = key.split("/")[0].split("_")[0];
-
-	const value = (val) => val == "null" ? undefined : val;
+	const key = params.key;
 	const type = util.getTypeByPath(key);
 	let checked = QINIU_AUDIT_STATE_NO_AUDIT;
 	if (type == "images") {
@@ -229,14 +166,11 @@ Files.prototype.qiniu = async function(ctx) {
 	}
 
 	let data = await this.model.upsert({
-		username: username,
 		type: type,
 		checked: checked,
-		key:params.key,
+		key: key,
 		hash: params.hash,
 		size: params.size,
-		sitename: value(params.sitename),
-		filename: value(params.filename),
 	});
 	
 	// 添加记录失败 应删除文件
@@ -332,30 +266,35 @@ Files.prototype.importOldData = async function(ctx) {
 	const params = ctx.state.params;
 	const self = this;
 
-	const apiUrl = params.apiUrl || (config.keepworkBaseURL + 'tabledb/query');
-	const data = {
-		tableName: "qiniu_files",
-		query: {},
-		pageSize: 1000000,
-	}
+	//const apiUrl = params.apiUrl || (config.keepworkBaseURL + 'tabledb/query');
+	//const data = {
+		//tableName: "qiniu_files",
+		//query: {},
+		//pageSize: 1000000,
+	//}
 
-	const result = await axios.post(apiUrl, data, {headers:	{"Authorization":"Bearer eyJhbGciOiJNRDUiLCJ0eXAiOiJKV1QifQ.eyJyb2xlSWQiOjEwLCJleHAiOjE1Mjk3NTY2MTgsImlzQWRtaW4iOnRydWUsInVzZXJJZCI6MTIsInVzZXJuYW1lIjoieGlhb3lhbyJ9.TkRobVlUTmhNV1ZqTkRFd1ltUTJOamM1T0RneE9XTXlaRE0yTVRoall6az0"}}).then(res => res.data);
+	//const result = await axios.post(apiUrl, data, {headers:	{"Authorization":"Bearer eyJhbGciOiJNRDUiLCJ0eXAiOiJKV1QifQ.eyJyb2xlSWQiOjEwLCJleHAiOjE1Mjk3NTY2MTgsImlzQWRtaW4iOnRydWUsInVzZXJJZCI6MTIsInVzZXJuYW1lIjoieGlhb3lhbyJ9.TkRobVlUTmhNV1ZqTkRFd1ltUTJOamM1T0RneE9XTXlaRE0yTVRoall6az0"}}).then(res => res.data);
 	
-	const list = result.data.data || [];
+	const apiUrl = params.apiUrl || (config.keepworkBaseURL + 'qiniu_files/export');
+	const list = await axios.get(apiUrl).then(res => res.data);
 	for (let i = 0; i < list.length; i++){
 		let item = list[i];
+
+		if (!item.userId) {
+			//console.log(item);
+			continue;
+		}
 		await self.model.upsert({
 			id: item._id,
 			key: item.key,
-		    username: item.username,
+		    userId: item.userId,
 			filename: item.filename,
 			checked: item.checked,
 			type: util.getTypeByPath(item.key),
 			size: item.size,
 		});
 	}
-	console.log(result);
-	return result;
+	return list;
 }
 
 Files.prototype.test = async function(ctx) {
@@ -411,6 +350,11 @@ Files.getRoutes = function() {
 		path: "qiniu",
 		method: "post",
 		action: "qiniu",
+		validate: {
+			body: {
+				key: joi.string().required(),
+			}
+		},
 	},
 	{
 		path: "audit",
@@ -451,14 +395,14 @@ Files.getRoutes = function() {
 		}
 	},
 	{
-		path: ":id",
+		path: "",
 		method: "POST",
 		action: "upsert",
 		authentated: true,
 		validate: {
 			body: {
 				key: joi.string().required(),
-			},
+			}
 		},
 	},
 	{
@@ -468,24 +412,24 @@ Files.getRoutes = function() {
 		authentated: true,
 		validate: {
 			params: {
-				id: joi.string().required(),
-			}
-		},
-	},
-	{
-		path: ":id/rename",
-		method: "PUT",
-		action: "update",
-		authentated: true,
-		validate: {
-			body: {
-				filename: joi.string().required(),
-			},
-			params: {
 				id: joi.number().required(),
 			}
 		},
 	},
+	//{
+		//path: ":id/rename",
+		//method: "PUT",
+		//action: "update",
+		//authentated: true,
+		//validate: {
+			//body: {
+				//filename: joi.string().required(),
+			//},
+			//params: {
+				//id: joi.number().required(),
+			//}
+		//},
+	//},
 	{
 		path: ":id",
 		method: "GET",
@@ -493,7 +437,7 @@ Files.getRoutes = function() {
 		authentated: true,
 		validate: {
 			params: {
-				id: joi.string().required(),
+				id: joi.number().required(),
 			}
 		},
 	},
